reafactor: 페이지 이름 변경 router 사용함수 수정
