+++ src/pages/LoginPage.js
@@ -1,4 +1,4 @@
-const Login = () => {
+const LoginPage = () => {
   return `
 <main class="bg-gray-100 flex items-center justify-center min-h-screen">
   <div class="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
@@ -24,4 +24,4 @@ const Login = () => {
 </main>`;
 };
 
-export default Login;
+export default LoginPage;
