+++ src/pages/NotFoundPage.js
@@ -1,4 +1,4 @@
-const NotFound = () => `
+const NotFoundPage = () => `
 <main class="bg-gray-100 flex items-center justify-center min-h-screen">
   <div class="bg-white p-8 rounded-lg shadow-md w-full text-center" style="max-width: 480px">
     <h1 class="text-2xl font-bold text-blue-600 mb-4">항해플러스</h1>
@@ -13,4 +13,4 @@ const NotFound = () => `
   </div>
 </main>`;
 
-export default NotFound;
+export default NotFoundPage;
